修改告警问题

+++ services/data_process/src/pipeline_node/multimedia_codec/encoder/encode_data_process.cpp
@@ -503,8 +503,8 @@ int32_t EncodeDataProcess::GetEncoderOutputBuffer(uint32_t index, Media::AVCodec
     int64_t timeStamp = info.presentationTimeUs;
     struct timespec time = {0, 0};
     clock_gettime(CLOCK_MONOTONIC, &time);
-    int64_t timeNs = static_cast<uint64_t>(time.tv_sec) * S2NS + static_cast<uint64_t>(time.tv_nsec);
-    int64_t encodeT = timeNs / US2NS - timeStamp;
+    int64_t timeNs = static_cast<int64_t>(time.tv_sec) * S2NS + static_cast<int64_t>(time.tv_nsec);
+    int64_t encodeT = timeNs / static_cast<int64_t>(US2NS) - timeStamp;
     int64_t finishEncodeT = GetNowTimeStampUs();
     int64_t startEncodeT = finishEncodeT - encodeT;
     bufferOutput->SetInt64(START_ENCODE_TIME_US, startEncodeT);
